+++ components/Forms/AppFormField.js
@@ -1,0 +1,26 @@
+import React from 'react';
+import { useFormikContext } from 'formik';
+
+import AppTextInput from './AppTextInput';
+import ErrorMessage from './ErrorMessage';
+
+function AppFormField({ name, ...otherProps  }) {
+    const { setFieldTouched, handleChange, errors, touched } = useFormikContext();
+
+    return (
+        <>
+            <AppTextInput
+                onBlur={() => setFieldTouched(name)}
+                onChangeText={handleChange(name)}
+                {...otherProps}
+            />
+            <ErrorMessage 
+                error={errors[name]}//Reuse name, don't need a error prop
+                visible={touched[name]}
+            />
+        </>
+    );
+}
+
+
+export default AppFormField;
